+++ src/extension/utils/createConfFile.ts
@@ -4,7 +4,13 @@
 
 import { workspace, Uri, window } from 'vscode'
 import { log, Sources } from './log'
-import { InputFormData, NewForm, SolcArg, SolidityObj } from '../types'
+import {
+  CONF_DIRECTORY_NAME,
+  InputFormData,
+  NewForm,
+  SolcArg,
+  SolidityObj,
+} from '../types'
 
 type ConfFile = {
   files?: string[]
@@ -437,7 +443,11 @@ export async function createConfFile(formData: InputFormData): Promise<void> {
     const encoder = new TextEncoder()
     const convertedData = convertSourceFormDataToConfFileJSON(formData)
     const content = encoder.encode(convertedData)
-    const path = Uri.joinPath(basePath.uri, 'conf', `${formData.name}.conf`)
+    const path = Uri.joinPath(
+      basePath.uri,
+      CONF_DIRECTORY_NAME,
+      `${formData.name}.conf`,
+    )
     await workspace.fs.writeFile(path, content)
     log({
       action: `Conf file was created`,