+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to this project will be documented in this file.
 
-### 0.0.5 (2022-12-11)
+### 0.0.5 (2023-01-02)
 
 ## Bug Fixes
 
@@ -10,10 +10,15 @@ All notable changes to this project will be documented in this file.
 * UI improvements in the settings form & results view ([#71] (https://github.com/Certora/vscode-certora-prover/pull/71))
 * Additional files always shows solc version, if it was provided ([#71] (https://github.com/Certora/vscode-certora-prover/pull/71))
 * Rename mode is fixed ([#71] (https://github.com/Certora/vscode-certora-prover/pull/71))
+* Verification report button appears when running on staging as well ([#72]((https://github.com/Certora/vscode-certora-prover/pull/72)))
 
 ## Features
 
+* Create jobs from conf files in certora/conf folder ([#72](https://github.com/Certora/vscode-certora-prover/pull/72))
 * Modal that asks the user if they are sure before deleting a job list item ([#71] (https://github.com/Certora/vscode-certora-prover/pull/71))
+* Changed location of conf files to certora/conf directory, and log files to .certora_internal directory ([#72](https://github.com/Certora/vscode-certora-prover/pull/72))
+* When clicking on the icon by a rule, create a new job with this one rule ([#76](https://github.com/Certora/vscode-certora-prover/pull/76))
+* 'Staging' option moved to the end of 'Additional Prover Settings' in the Settings Form ([#76](https://github.com/Certora/vscode-certora-prover/pull/76))
 
 ### 0.0.4 (2022-11-23)